+++ src/components/App/index.tsx
@@ -20,7 +20,7 @@ const App = (): JSX.Element => {
     <Style className='app' onClick={handleClick}>
       <Context.Provider value={{ dispatch, state }}>
         <PlayArea activeCard={activeCard} setActiveCard={setActiveCard} />
-        <ArrowIcon />
+        <ArrowIcon iconType='END_TURN' />
         <HandIcon />
         <Hand />
       </Context.Provider>