+++ src/components/ArrowIcon/style.tsx
@@ -2,10 +2,13 @@ import styled from 'styled-components';
 
 const Style = styled.div`
   cursor: pointer;
-  display: inline-block;
-  position: fixed;
-  right: 22px;
-  top: 22px;
+
+  &.endTurnIcon {
+    display: inline-block;
+    position: fixed;
+    right: 22px;
+    top: 22px;
+  }
 `;
 
 export default Style;